+++ src/lib/Endpoint/Mysql.php
@@ -104,7 +104,7 @@ class Mysql extends AbstractSqlDatabase
 
         while ($row = $result->fetch_assoc()) {
             yield $this->build($row);
-            $i;
+            ++$i;
         }
 
         return $i;